Реализовано: - Регистрация мока сетевого клиента как часть DIFramework; - Внедрение Мока сетевого клиента в DIContainer Example таргета модуля Characters
